Rename napari_btrack.track to napari_btrack.main
