+++ napari_btrack/napari.yaml
@@ -4,7 +4,7 @@ contributions:
   commands:
     - id: napari-btrack.track
       title: Create Track
-      python_name: napari_btrack.track:track
+      python_name: napari_btrack.main:track
   widgets:
     - command: napari-btrack.track
       display_name: Track
